Replace get_all_length_types because it is no longer correct
This stems from the commit b2ee5a. Because the order of the outer sets
does matter with respect to the inner, this becomes a cartesian product
instead of the mess that it was before. There is some trimming that can
be done with respect to decreasing the number of inner groups, but as a
whole it will not change much.