Add function to remove columns with only zeros
This should dramatically reduce memory usage and running time.
Really, this change is not a good thing to do, but practically
speaking it should help tremendously.

The issue with doing this sort of method is that it does not allow
expansion outside the sets that were used (though, practically
speaking they would have to be in the training data anyways).