Add fix to KRR when using GridSearchCV
The standard GridSearchCV code in scikit-learn does not init a new clf
every single time. Instead, they clone the old one and then use
set_params. This means that the old code would throw an error if one
tried to use a laplace kernel because the laplace kernel does not exist
by default in scikit learn.

This fix basically replaces the __init__ code with code that makes
calls to set_params instead.